pyproject.toml
[project]
name = "data-flows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ephem>=4.2",
    "ipykernel>=6.29.5",
    "opencv-python>=4.10.0",
    "pandas>=2.3.1",
    "pillow>=11.0.0",
    "pip>=25.1.1",
    "prefect>=3.4.6",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
]

[project.optional-dependencies]
opencv = ["opencv-python>=4.10.0"]

[tool.setuptools.packages.find]
where = ["src"]
